backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.3
websockets>=10.4